质数算法的快慢速度重测.

+++ src/common/common.java
@@ -15,13 +15,13 @@ public class common {
         return array;
     }
 
-    public static boolean isPrime1(int candidate) {
-        /*
-         * 较慢的判断质数算法
+    public static boolean isPrime1(long candidate) {
+        /**
+         * 较快的判断质数算法
          * */
         if ((candidate & 1) != 0) // 是奇数
         {
-            int limit = (int)Math.sqrt(candidate);
+            long limit = (long)Math.sqrt(candidate);
             for (int divisor = 3; divisor <= limit; divisor += 2) // divisor = 3、5、7...candidate的平方根
             {
                 if ((candidate % divisor) == 0)
@@ -32,9 +32,9 @@ public class common {
         return (candidate == 2); // 除了2，其它偶数全都不是质数
     }
 
-    public static boolean isPrime2(int num) {
-        /*
-        * 更快的判断质数算法
+    public static boolean isPrime2(long num) {
+        /**
+        * 较慢的判断质数算法
         * */
         if (num <= 3) {
             return num > 1;
@@ -43,7 +43,7 @@ public class common {
         if (num % 6 != 1 && num % 6 != 5) {
             return false;
         }
-        int sqrt = (int) Math.sqrt(num);
+        long sqrt = (long) Math.sqrt(num);
         for (int i = 5; i <= sqrt; i += 6) {
             if (num % i == 0 || num % (i + 2) == 0) {
                 return false;
